// types/responses/vocabulary.response.ts
export interface VocabularyResponse {
    id: number;
    japanese: string;
    reading?: string;
    vietnameseMeaningText?: string;
    englishMeaningText?: string;
}

export interface GrammarResponse {
    id: number;
    japanese: string;
    reading?: string;
    vietnameseMeaningText?: string;
    englishMeaningText?: string;
}